Cel Shading, Blue & Red Gradient Textures, and Post Processsing
Cel Shading, Blue and Red materials were made. Just transfer over exactly the post processing volume box's settings in ShaderTestingAreaV2 to the main map and the comuter should do the work for you. Then if you want something red just attatch the Red material to it or Blue if you want it Blue.